+++ GameFramework/Include/Object/Bullet.h
@@ -14,7 +14,7 @@ protected:
 	EBullet_Type m_BulletType;
 	Vector2		m_Dir;
 	float		m_Distance;
-	float		m_Damage;
+	int			m_Damage;
 	CGun*		m_Owner;
 
 public:
@@ -23,7 +23,7 @@ public:
 		m_Owner = Owner;
 	}
 	void SetBulletType(EBullet_Type BType);
-	void SetBulletDamage(float Damage)
+	void SetBulletDamage(int Damage)
 	{
 		m_Damage = Damage;
 	}
